[Refactor] Convert to Repository Pattern
